requirements.txt
absl-py==2.1.0
alembic==1.14.0
asttokens==3.0.0
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.40.0
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.15
docker==7.1.0
executing==2.2.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.2
gast==0.6.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.68.1
h5py==3.12.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.7
libclang==18.1.1
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mlflow==2.19.0
mlflow-skinny==2.19.0
mysql-connector-python==9.1.0
namex==0.0.8
narwhals==1.16.0
nest-asyncio==1.6.0
numpy==1.26.4
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.6.0
scipy==1.13.1
setuptools==75.6.0
six==1.17.0
smart-open==7.1.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
zipp==3.21.0
